// src/_model/Type.ts
export enum Type {
  Normal = 'normal',
  Grass = 'grass',
  Poison = 'poison',
  Fire = 'fire',
  Water = 'water',
  Electric = 'electric',
  Ice = 'ice',
  Fighting = 'fighting',
  Ground = 'ground',
  Flying = 'flying',
  Psychic = 'psychic',
  Bug = 'bug',
  Rock = 'rock',
  Ghost = 'ghost',
  Dark = 'dark',
  Dragon = 'dragon',
  Steel = 'steel',
  Fairy = 'fairy',
}